pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "nobiart", email = "dmitri.besleaga@gmail.com" }
]
dependencies = [
    "websocket-client>=1.8.0",
    "requests>=2.32.3",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]
